+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for import from Iceberg table definitions.
+
 ## [0.10.13] - 2024-09-20
 
 ### Added